Report_Monaco_Racing/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dominate==2.7.0
flasgger==0.9.5
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
mistune==2.0.4
packaging==21.3
peewee==3.15.3
pluggy==1.0.0
prettytable==3.4.1
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-flask==1.2.0
python-simplexml==0.1.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
xmltodict==0.13.0
